fix: 장바구니 상태 요청 selector와 상태값 저장 atom 사이의 순환 참조 문제 해결
