+++ src/store/atom/atoms.ts
@@ -1,20 +1,7 @@
 import { atomFamily, atom, selector } from "recoil";
 import { LOCAL_STORAGE_KEY } from "../../constants";
-import { fetchProducts } from "../api";
 import { getStorage, setStorage } from "../localStorage/localStorage";
-
-export const fetchCartState = selector({
-  key: "fetchCartState",
-  get: async () => {
-    const { content }: { content: CartItemInfo[] } = await fetchProducts("GET");
-    const localData = getStorage<CartItemCheckedStateInStorage>(LOCAL_STORAGE_KEY, {});
-    content.forEach((cartItem) => {
-      if (localData[cartItem.id] === undefined) localData[cartItem.id] = true;
-    });
-    setStorage(LOCAL_STORAGE_KEY, localData);
-    return content;
-  },
-});
+import { fetchCartState } from "../selector/fetchCartState";
 
 export const cartState = atom({
   key: "cartState",
